feat(zappy_net_ringbuf): add flush and read from FD operations for ring buffer
