+++ libzappy_net/Makefile
@@ -39,7 +39,8 @@ SRC_C = $(SRC_DIR)/libzappy_net.c \
         $(SRC_DIR)/zappy_net_ringbuf_check.c \
         $(SRC_DIR)/zappy_net_ringbuf_write.c \
         $(SRC_DIR)/zappy_net_ringbuf_read.c \
-        $(SRC_DIR)/zappy_net_ringbuf_io.c \
+        $(SRC_DIR)/zappy_net_ringbuf_flush.c \
+        $(SRC_DIR)/zappy_net_ringbuf_read_fd.c \
         $(SRC_DIR)/zappy_net_io.c
 SRC_CPP =
